neuprint: elements: Ensure properties which represent a location are converted to neo4j spatial points and also don't emit 'nan' in the CSV
